+++ docs/changelog.md
@@ -13,6 +13,15 @@
   delimiter (CSV dialect detection)
 - Export CSV in batches
 - Change CSV dialect detection sample size to 256KiB
+- [#242](https://github.com/turicas/rows/issues/242) (CLI) Add
+  `--fields`/`--fields-exclude` to `convert`, `join` and `sum` (and rename
+  `--fields-exclude` on `print`), also remove `--fields` from `query` (is not
+  needed).
+- [#235](https://github.com/turicas/rows/issues/235) Implement `--http-cache`
+  and `--http-cache-path`.
+- [#237](https://github.com/turicas/rows/issues/237) Implement `rows schema`
+  (generates schema in text, SQL and Django models).
+
 
 ### Bug Fixes